backend-django/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
